+++ gulpfile.js
@@ -33,7 +33,7 @@ gulp.task('build', function() {
     .pipe(gulp.dest(config.buildDir));
 });
 
-gulp.task('watch', function () {
+gulp.task('watch', ['build'], function () {
   gulp.watch('./src/**/*.js', ['build']);
 });