+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   + they can add more than 1 variable in the environment
     * specify `data-binding="*"` in the query tag, then return an object
   + a `form()` helper allows to call the API and get a Q promise
+  + an `ajax()` helper allows to perform HTTP(S) requests and get a Q promise
 
 ## 0.1.0